+++ src/classes/index.ts
@@ -1,0 +1,5 @@
+export { default as GatewayStitcher } from './GatewayStitcher';
+export { default as GraphQLExecutor } from './GraphQLExecutor';
+export { default as RequestHandler, Request } from './RequestHandler';
+export { default as requestParser } from './RequestParser';
+export { default as SchemaBuilder } from './SchemaBuilder';
