+++ src/mixins/gatewayMixin.ts
@@ -1,8 +1,7 @@
 import type { ServerResponse } from 'http';
 import type { Service, ServiceSchema } from 'moleculer';
-import GatewayStitcher from '../GatewayStitcher';
-import RequestHandler from '../RequestHandler';
-import type { Request } from '../RequestHandler';
+import { GatewayStitcher, RequestHandler } from '../classes';
+import type { Request } from '../classes';
 
 interface GatewayService extends Service {
 	rebuildSchema: boolean;
